refactor: input type 정의를 auth에서 common으로 이동
