+++ types/Common.ts
@@ -1,0 +1,9 @@
+export interface AuthInputProps {
+  type: string;
+  id: string;
+  value: string;
+  onChange: (e: React.ChangeEvent<HTMLInputElement>) => void;
+  placeholder: string;
+  error?: boolean;
+  message?: string;
+}
